# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel API",
        "description": "Sayohat veb-sayti uchun API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://toqtarbay.dbc.uz/api",
            "description": "Production Server"
        },
        {
            "url": "http://travel.test/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/tours": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Barcha turlarni olish",
                "description": "Aktiv turlarni ro'yxatini qaytaradi. Category bo'yicha filter qilish mumkin.",
                "operationId": "fee1453175a1fa9b6f205a3265be39eb",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "uz",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Kategoriya ID (ixtiyoriy)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli javob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Orol dengizi safari"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Ajoyib sayohat..."
                                                    },
                                                    "routes": {
                                                        "type": "string",
                                                        "example": "Toshkent - Nukus - Moynaq"
                                                    },
                                                    "important_info": {
                                                        "type": "string",
                                                        "example": "Pasport talab qilinadi"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 1500000
                                                    },
                                                    "duration_days": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "duration_nights": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "min_age": {
                                                        "type": "integer",
                                                        "example": 18
                                                    },
                                                    "max_people": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "rating": {
                                                        "type": "number",
                                                        "example": 4.5
                                                    },
                                                    "reviews_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tabiiy safari"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "main_image": {
                                                        "type": "string",
                                                        "example": "/storage/uploads/image.jpg"
                                                    },
                                                    "images": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "url": {
                                                                    "type": "string",
                                                                    "example": "/storage/uploads/image.jpg"
                                                                },
                                                                "is_main": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}